backend/web/requirements.tests.txt
sqlalchemy_utils>=0.34
yarl>=1.3
pytest>=4.6
pytest-flask>=0.15.0
